backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
celery[redis]==5.4.0
pymongo==4.10.1
redis==5.1.1
selenium==4.26.1
python-dotenv==1.0.1
langchain==0.3.3
langchain-openai==0.2.2
langchain-community==0.3.2
faiss-cpu==1.9.0
numpy==1.26.4
requests==2.32.3